Refactor 'news-article' til 'featured-article'
