+++ src/components/_common/headers/featured-header/FeaturedHeader.module.scss
@@ -1,9 +1,6 @@
 @use 'src/common' as common;
-// @use '../../../pages/situation-page/SituationPage' as SituationPage;
-// @use '../../../pages/product-page/ProductPage' as ProductPage;
-// @use '../../../pages/guide-page/GuidePage' as GuidePage;
 
-.themedPageHeader {
+.featuredHeader {
     display: flex;
     justify-content: center;
     background-color: white;
