+++ src/components/pages/featured-article-page/FeaturedArticlePage.module.scss
@@ -2,7 +2,7 @@
 
 $margin: 2rem;
 
-.newsArticlePage {
+.featuredArticlePage {
     display: flex;
     flex-direction: column;
